# sdk/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ChainSettle API",
    "description": "Backend API for ChainSettle — milestone-based supply chain escrow on Stellar Soroban. Regenerate with `npm run generate:sdk`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Envelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {},
          "timestamp": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": ["users"],
        "operationId": "UsersController_getProfile",
        "summary": "Get the authenticated user profile",
        "security": [{ "bearer": [] }],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Envelope" }
              }
            }
          }
        }
      }
    }
  }
}